+++ spec/rails_helper.rb
@@ -42,7 +42,7 @@ RSpec.configure do |config|
   end
 
   FactoryBot.definition_file_paths = [
-    File.expand_path("factories", __dir__)
+    File.expand_path("factories", __dir__),
   ]
   FactoryBot.find_definitions