+++ CHANGELOG.md
@@ -56,3 +56,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add Keys Transformer
 - Always transform request keys (deep transform) to snake_case
 - Configurable: transform response keys to either: snake_case, lowerCamelCase, or UpperCamelCase
+- Implement request rate limiting using Rack::Attack